# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Modular API server. Run in platform or tenant mode via -mode flag.",
        "title": "Go Web API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tenants": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new tenant with optional custom database URL. Creator becomes owner automatically.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenants"
                ],
                "summary": "Create a new tenant (authenticated users)",
                "parameters": [
                    {
                        "description": "Tenant data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_tenancy_models.CreateTenantDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_tenancy_models.TenantDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tenants/my-tenants": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns list of tenants that the authenticated user can access",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenants"
                ],
                "summary": "Get user's accessible tenants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/josex_web_modules_tenancy_models.UserTenantResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tenants/self-service": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Allow authenticated users to create tenants with limits based on their subscription plan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenants"
                ],
                "summary": "Create tenant (self-service with subscription limits)",
                "parameters": [
                    {
                        "description": "Tenant data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_tenancy_models.CreateTenantDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_tenancy_models.TenantDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_slug}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update tenant name, database URL, schema, or settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenants"
                ],
                "summary": "Update tenant information (owner/admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant slug",
                        "name": "tenant_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated tenant data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_tenancy_models.UpdateTenantDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_tenancy_models.TenantDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_slug}/migrate": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Execute all pending migrations on a tenant's custom database (admin/super_admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenants"
                ],
                "summary": "Run migrations on tenant database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant slug",
                        "name": "tenant_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_slug}/users": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a user to the tenant with specified role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenants"
                ],
                "summary": "Add a user to tenant (owner/admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant slug",
                        "name": "tenant_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User and role data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_tenancy_models.AddUserToTenantDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_slug}/users/{user_id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove a user from the tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenants"
                ],
                "summary": "Remove a user from tenant (owner/admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant slug",
                        "name": "tenant_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_slug}/users/{user_id}/role": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Change the role of a tenant member. Only owner can assign the owner role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenants"
                ],
                "summary": "Update a user's role in the tenant (owner/admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant slug",
                        "name": "tenant_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New role",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_tenancy_models.UpdateUserRoleDto"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "josex_web_modules_core_errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code\nexample: 400",
                    "type": "string"
                },
                "detail": {
                    "description": "Error details\nexample: {\"email\": \"Invalid email format\"}"
                },
                "error": {
                    "description": "Error message\nexample: Bad Request"
                },
                "message": {
                    "description": "Translated error message"
                }
            }
        },
...
        "josex_web_modules_tenancy_models.AddUserToTenantDto": {
            "type": "object",
            "required": [
                "role",
                "user_id"
            ],
            "properties": {
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "josex_web_modules_tenancy_models.CreateTenantDto": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "database_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "schema_name": {
                    "type": "string"
                },
                "settings": {
                    "description": "JSONB as string",
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "josex_web_modules_tenancy_models.TenantDetailResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_suspended": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "schema_name": {
                    "type": "string"
                },
                "settings": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "josex_web_modules_tenancy_models.UpdateTenantDto": {
            "type": "object",
            "properties": {
                "database_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "schema_name": {
                    "type": "string"
                },
                "settings": {
                    "description": "JSONB as string",
                    "type": "string"
                }
            }
        },
        "josex_web_modules_tenancy_models.UpdateUserRoleDto": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "role": {
                    "type": "string",
                    "enum": [
                        "owner",
                        "admin",
                        "member"
                    ]
                }
            }
        },
        "josex_web_modules_tenancy_models.UserTenantResponse": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "is_suspended": {
                    "type": "boolean"
                },
                "joined_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "user_role": {
                    "type": "string"
                }
            }
        },
...
    }
}